redesign literals to work on gpu mode
