+++ mcdc/code_factory/numba_objects_generator.py
@@ -1187,49 +1187,3 @@ def singular_to_plural(word: str) -> str:
         parts[-1] = w + "s"
 
     return "_".join(parts)
-
-
-# ==============================================================================
-# MC/DC Member Array Sizes
-# ==============================================================================
-
-
-def literalize(value):
-    jit_str = f"@njit\ndef impl():\n    return {value}\n"
-    exec(jit_str, globals(), locals())
-    return eval("impl")
-
-
-def rpn_buffer_size():
-    pass
-
-
-def make_size_rpn(cells):
-    global rpn_buffer_size
-    size = max([np.sum(np.array(x.region_RPN_tokens) >= 0.0) for x in cells])
-    rpn_buffer_size = literalize(size)
-
-
-# ======================================================================================
-# Make literals
-# ======================================================================================
-
-
-def make_literals(simulation):
-    # Sizes
-    if len(simulation.cells) == 0:
-        rpn_evaluation_buffer_size = 1
-    else:
-        rpn_evaluation_buffer_size = int(
-            max(
-                [np.sum(np.array(x.region_RPN_tokens) >= 0.0) for x in simulation.cells]
-            )
-        )
-
-    path = f"{Path(mcdc.__file__).parent}"
-    with open(f"{path}/transport/literals.py", "w") as f:
-        text = "# The following is automatically generated by code_factory.py\n\n"
-
-        text += f"rpn_evaluation_buffer_size = {rpn_evaluation_buffer_size}\n"
-
-        f.write(text)
