--- mcdc/transport/.gitignore
@@ -1,1 +1,0 @@
-literals.py
