+++ mcdc/transport/__init__.py
@@ -1,4 +1,5 @@
 import mcdc.transport.geometry as geometry
+import mcdc.transport.literals as literals
 import mcdc.transport.particle_bank as particle_bank
 import mcdc.transport.simulation as simulation
 import mcdc.transport.util as util
