+++ mcdc/transport/literals.py
@@ -1,0 +1,6 @@
+# The following will be replaced with their respective literals by
+# code_factory/literals_generator.py
+
+
+def rpn_evaluation_buffer_size():
+    pass
